Login + Registration

+++ lib/features/onboarding/presentation/widgets/onboarding_screen_layout.dart
@@ -34,15 +34,18 @@ class OnboardingScreenLayout extends StatelessWidget {
             final screenWidth = constraints.maxWidth;
             final screenHeight = constraints.maxHeight;
             final horizontalPadding =
-                (screenWidth * 0.075).clamp(24.0, 30.0);
+                (screenWidth * 0.075).clamp(24.0, 30.0).toDouble();
             final illustrationWidth =
-                (screenWidth * 0.78).clamp(260.0, 320.0);
+                (screenWidth * 0.78).clamp(260.0, 320.0).toDouble();
             final titleSize = screenWidth < 360 ? 38.0 : 41.0;
-            final topSpacing = (screenHeight * 0.035).clamp(18.0, 30.0);
-            final titleSpacing = (screenHeight * 0.042).clamp(26.0, 38.0);
+            final topSpacing =
+                (screenHeight * 0.035).clamp(18.0, 30.0).toDouble();
+            final titleSpacing =
+                (screenHeight * 0.042).clamp(26.0, 38.0).toDouble();
             final indicatorSpacing =
-                (screenHeight * 0.038).clamp(22.0, 32.0);
-            final bottomSpacing = (screenHeight * 0.04).clamp(22.0, 34.0);
+                (screenHeight * 0.038).clamp(22.0, 32.0).toDouble();
+            final bottomSpacing =
+                (screenHeight * 0.04).clamp(22.0, 34.0).toDouble();
 
             return Padding(
               padding: EdgeInsets.fromLTRB(
